Add utils.py to CANviewer and allow application to exit gracefully
