+++ projects/canviewer/canserver.py
@@ -5,17 +5,20 @@ import threading
 db = None
 
 
-def listener_fn(can_bus, callback):
+def listener_fn(can_bus, callback, kill_flag):
     """Thread that runs all the time to listen to CAN messages
 
     References:
       - https://python-can.readthedocs.io/en/master/interfaces/socketcan.html
       - https://python-can.readthedocs.io/en/master/
     """
-    while True:
+    while not kill_flag.is_set():
         msg = can_bus.recv(1)  # 1 second receive timeout
         if msg:
             callback(msg, db)
+
+    can_bus.shutdown()
+    print('Exited gracefully')
 
 
 def init_can(channel, bustype, bitrate, rx_callback, dbc):
@@ -26,15 +29,15 @@ def init_can(channel, bustype, bitrate, rx_callback, dbc):
         bitrate=bitrate,
     )
 
+    kill_flag = threading.Event()
     listener = threading.Thread(
         target=listener_fn,
         name="listener",
-        kwargs={"can_bus": can_bus, "callback": rx_callback},
+        kwargs={"can_bus": can_bus, "callback": rx_callback, "kill_flag": kill_flag},
     )
 
-    listener.daemon = True
     listener.start()
 
     db = cantools.database.load_file(dbc)
 
-    return can_bus, db
+    return can_bus, db, kill_flag
